Update hiding.md
A few proposed changes here.  First, it appears "authentication" was used before it was defined--I think because it was meant to be "authorization".

Second, security is much more than access control.  Authentication and Authorization generally combine in the security literature for the purpose of access control, which seems to be the focus and link to privacy above.

Not changed, but what I would suggest, would be to reverse the order. You can talk about authenticating an identity/party (person/group/developer/application/etc) in the absence of used that identity to authorize access (a noun without a verb), but you can't really talk about the action authorizing an identity/party to do something (the verb) without that noun.  This is why the security literature typically puts authentication before authorization.